pyproject.toml
[tool.poetry]
name = "pokemon-bot"
version = "0.1.0"
description = "A Discord bot for Pokémon cards"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.5"
discord-py = "^2.6.3"
uuid = "^1.30"
pandasai = {extras = ["all"], version = "^2.3.2"}
numpy = "1.23.5" 
pyyaml = "^6.0" 
tabulate = "^0.9.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.9"}

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"